Move out feature matrix funcs in OutputTools.
Note that the feature matrix related code will not work for now.
